terminate basic #1

+++ run_terminate_imagebuilder_gendo_old_environment.py
@@ -5,6 +5,9 @@ import re
 from run_common import AWSCli
 from run_common import print_message
 from run_common import print_session
+from datetime import datetime
+from datetime import timedelta
+from env import env
 
 options, args = dict(), list()
 
@@ -164,10 +167,67 @@ def run_terminate_image(name):
         else:
             abnormal_resource_version_list.append(tt)
 
+    print_message(f'normal resource version list : {normal_resource_version_list}')
+    print_message(f'abnormal resource version list : {abnormal_resource_version_list}')
+
     delete_version_list = list(abnormal_resource_version_list)
+
+    # 1. 두달 전 이여야 함.
+    # 2. 가진 것 중 최신 2개만 남김.
+    ec2_describe_role_arn = ''
+    used_ec2_ami = ''
+    for settings in env.get('imagebuilder', list()):
+        if settings['NAME'] == 'run_terminate_imagebuilder_gendo_old_environment':
+            ec2_describe_role_arn = settings['EC2_DESCRIBE_ROLE_ARN']
+
+    if ec2_describe_role_arn:
+        cmd = ['sts', 'assume-role']
+        cmd += ['--role-arn', ec2_describe_role_arn]
+        cmd += ['--role-session-name', 'ec2-describe-role']
+        rr = aws_cli.run(cmd)
+
+        access_key = rr['Credentials']['AccessKeyId']
+        secret_key = rr['Credentials']['SecretAccessKey']
+        session_token = rr['Credentials']['SessionToken']
+
+        aws_cli_for_ec2 = AWSCli(aws_access_key=access_key,
+                                aws_secret_access_key=secret_key,
+                                aws_session_token=session_token)
+
+        cmd = ['elasticbeanstalk', 'describe-environments']
+    'elasticbeanstalk describe-environments'
+
+        cmd = ['ec2', 'describe-instances']
+        cmd += ['--filters',f'Name=tag:elasticbeanstalk:environment-name,Values={gendo_elb_name}']
+        rr = aws_cli_for_ec2.run(cmd)
+        print(rr)
+
+        # used_ec2_ami =
+        # aws configure set profile.test3.aws_access_key_id "ASIAQJWZ32MVMEHY62SW"
+    # aws configure set profile.test3.aws_secret_access_key "3GKkrimxMbqQTKxK7N0h4oz/CqL14yjgb2rw29op"
+    # aws configure set profile.test3.aws_session_token "IQoJb3JpZ2luX2VjEHgaDmFwLW5vcnRoZWFzdC0yIkcwRQIgJ9ERMC9ImLEwur0aG8NTIcdW73NQoHMmM2ybh/7/uf8CIQDNV0SmbnZ0b/6PtYYLqR6/nz0mKqGwCkxHK5se/IYPJCqRAggxEAMaDDAyMDg1ODAwMDE3MCIMjfeVe7ZT1hfD7titKu4BiZLX9iYmGzzzIxHq/Rusm5vH2+MzaZkHapb4FJCbMFyrqeTnxdUcLIJqgwws7k/KLLvEvVUOaLhWco3zNTSo33ILdtC+piWXEIYAaA0HkMhiNeDiUa1Odj8Weq7g5+Fva+jHglEExaY62lKBnmYuNvYC5ZfsFZ/6IbAdipeRSjx/l+OP9lr5Y52lqz29EgBFTYxVhm6uJ55iuuHSoowefRwtEgv8ZfLgyoQ2/5DesNWwLSGo64uiwGwAuWQXfuqNJM7EWDpJ4V2gZDvUqrBigW7pCP30mEKhL/wdx9XlZcxT+Gk4w0uVy7snRUJsTDC01dGZBjqdAYtQ5BvW0T6juxfrk1nRZ3/LxTV7wrUOg8mV2fjjjwY47Hmesq0bmBWfsu/rcOwvnWaEz+rzbW5iJXknze1mZ9LYH0o39DNfAbNKGmyuvpAZ/so/YCyjqMAUAxHFkWP5sN7NlqpM/WSGhzeNclZAnEv4mSRoxVkAxz3BO/oajEjWchdLPsES216juxOyUqM8l4+TFVf1+A1Zg+YBX48="
+    # aws configure set profile.test3.region "ap-northeast-2"
+    # 'set region us-west-1 --profile testing'
+
+    # timestamp_8_weeks_ago = datetime.now() - timedelta(weeks=8)
+    # # describe_ec2
+    # deleted_target_normal_resource_version_list = list()
+    # for vv in normal_resource_version_list:
+    #
+    #
+    #     if int(vv) < int(timestamp_8_weeks_ago):
+    #         deleted_target_normal_resource_version_list.append(vv)
+
+
     if len(normal_resource_version_list) > 2:
         normal_resource_version_list.sort(key=int)
         delete_version_list += normal_resource_version_list[2:]
+
+    if not delete_version_list:
+        print_message('There are no versions to delete.')
+        return
+
+    print_message(f'delete version list : {delete_version_list}')
 
     for ami_arn in ami_arn_list:
         if delete_version_list_any_imagebuilder_resource(delete_version_list, ami_arn):
@@ -178,75 +238,75 @@ def run_terminate_image(name):
                 cmd = ['imagebuilder', 'list-image-build-versions']
                 cmd += ['--image-version-arn', ami_arn]
                 arn_version_list = aws_cli.run(cmd, ignore_error=True)
-
-                if arn_version_list['imageSummaryList']:
-                    cmd = ['imagebuilder', 'delete-image']
-                    cmd += ['--image-build-version-arn', r['arn']]
-                    aws_cli.run(cmd, ignore_error=True)
-
-    print_message(f'delete ec2 {name} ami and snapshot')
-    for img in ec2_gendo_img_list:
-        if delete_version_list_any_ec2_image(delete_version_list, img):
-            ami = img['ImageId']
-            snapshot_id = img['BlockDeviceMappings'][0]['Ebs']['SnapshotId']
-
-            cmd = ['ec2', 'deregister-image']
-            cmd += ['--image-id', ami]
-            aws_cli.run(cmd)
-
-            cmd = ['ec2', 'delete-snapshot']
-            cmd += ['--snapshot-id', snapshot_id]
-            aws_cli.run(cmd)
-
-    print_message(f'delete imagebuilder {name} pipe lines')
-    for pipe_line in pipe_line_list:
-        if delete_version_list_any_imagebuilder_resource(delete_version_list, pipe_line):
-            cmd = ['imagebuilder', 'delete-image-pipeline']
-            cmd += ['--image-pipeline-arn', pipe_line]
-            aws_cli.run(cmd)
-
-    print_message(f'delete imagebuilder {name} distributions')
-    for distribution in distribution_list:
-        if delete_version_list_any_imagebuilder_resource(delete_version_list, distribution):
-            cmd = ['imagebuilder', 'delete-distribution-configuration']
-            cmd += ['--distribution-configuration-arn', distribution]
-            aws_cli.run(cmd)
-
-    print_message(f'delete imagebuilder {name} infrastructures')
-    for infrastructure in infrastructure_list:
-        if delete_version_list_any_imagebuilder_resource(delete_version_list, infrastructure):
-            cmd = ['imagebuilder', 'delete-infrastructure-configuration']
-            cmd += ['--infrastructure-configuration-arn', infrastructure]
-            aws_cli.run(cmd)
-
-    print_message(f'delete imagebuilder {name} image-recipes')
-    for image_recipe in image_recipe_list:
-        if delete_version_list_any_imagebuilder_resource(delete_version_list, image_recipe):
-            cmd = ['imagebuilder', 'delete-image-recipe']
-            cmd += ['--image-recipe-arn', image_recipe]
-            aws_cli.run(cmd)
-
-    print_message(f'delete imagebuilder {name} components')
-    for component in component_list:
-        if delete_version_list_any_imagebuilder_resource(delete_version_list, component):
-            cmd = ['imagebuilder', 'list-component-build-versions']
-            cmd += ['--component-version-arn', component]
-            rr = aws_cli.run(cmd)
-            for r in rr['componentSummaryList']:
-                cmd = ['imagebuilder', 'list-component-build-versions']
-                cmd += ['--component-version-arn', component]
-                arn_version_list = aws_cli.run(cmd, ignore_error=True)
-                if arn_version_list['componentSummaryList']:
-                    cmd = ['imagebuilder', 'delete-component']
-                    cmd += ['--component-build-version-arn', r['arn']]
-                    aws_cli.run(cmd, ignore_error=True)
-
-    print_message('delete cloudwatch image builder logs')
-    for log_group in imagebuilder_cw_log_list:
-        if delete_version_list_any_cloudwatch_log(delete_version_list, log_group):
-            cmd = ['logs', 'delete-log-group']
-            cmd += ['--log-group-name', log_group['logGroupName']]
-            aws_cli.run(cmd, ignore_error=True)
+    #
+    #             if arn_version_list['imageSummaryList']:
+    #                 cmd = ['imagebuilder', 'delete-image']
+    #                 cmd += ['--image-build-version-arn', r['arn']]
+    #                 aws_cli.run(cmd, ignore_error=True)
+    #
+    # print_message(f'delete ec2 {name} ami and snapshot')
+    # for img in ec2_gendo_img_list:
+    #     if delete_version_list_any_ec2_image(delete_version_list, img):
+    #         ami = img['ImageId']
+    #         snapshot_id = img['BlockDeviceMappings'][0]['Ebs']['SnapshotId']
+    #
+    #         cmd = ['ec2', 'deregister-image']
+    #         cmd += ['--image-id', ami]
+    #         aws_cli.run(cmd)
+    #
+    #         cmd = ['ec2', 'delete-snapshot']
+    #         cmd += ['--snapshot-id', snapshot_id]
+    #         aws_cli.run(cmd)
+    #
+    # print_message(f'delete imagebuilder {name} pipe lines')
+    # for pipe_line in pipe_line_list:
+    #     if delete_version_list_any_imagebuilder_resource(delete_version_list, pipe_line):
+    #         cmd = ['imagebuilder', 'delete-image-pipeline']
+    #         cmd += ['--image-pipeline-arn', pipe_line]
+    #         aws_cli.run(cmd)
+    #
+    # print_message(f'delete imagebuilder {name} distributions')
+    # for distribution in distribution_list:
+    #     if delete_version_list_any_imagebuilder_resource(delete_version_list, distribution):
+    #         cmd = ['imagebuilder', 'delete-distribution-configuration']
+    #         cmd += ['--distribution-configuration-arn', distribution]
+    #         aws_cli.run(cmd)
+    #
+    # print_message(f'delete imagebuilder {name} infrastructures')
+    # for infrastructure in infrastructure_list:
+    #     if delete_version_list_any_imagebuilder_resource(delete_version_list, infrastructure):
+    #         cmd = ['imagebuilder', 'delete-infrastructure-configuration']
+    #         cmd += ['--infrastructure-configuration-arn', infrastructure]
+    #         aws_cli.run(cmd)
+    #
+    # print_message(f'delete imagebuilder {name} image-recipes')
+    # for image_recipe in image_recipe_list:
+    #     if delete_version_list_any_imagebuilder_resource(delete_version_list, image_recipe):
+    #         cmd = ['imagebuilder', 'delete-image-recipe']
+    #         cmd += ['--image-recipe-arn', image_recipe]
+    #         aws_cli.run(cmd)
+    #
+    # print_message(f'delete imagebuilder {name} components')
+    # for component in component_list:
+    #     if delete_version_list_any_imagebuilder_resource(delete_version_list, component):
+    #         cmd = ['imagebuilder', 'list-component-build-versions']
+    #         cmd += ['--component-version-arn', component]
+    #         rr = aws_cli.run(cmd)
+    #         for r in rr['componentSummaryList']:
+    #             cmd = ['imagebuilder', 'list-component-build-versions']
+    #             cmd += ['--component-version-arn', component]
+    #             arn_version_list = aws_cli.run(cmd, ignore_error=True)
+    #             if arn_version_list['componentSummaryList']:
+    #                 cmd = ['imagebuilder', 'delete-component']
+    #                 cmd += ['--component-build-version-arn', r['arn']]
+    #                 aws_cli.run(cmd, ignore_error=True)
+    #
+    # print_message('delete cloudwatch image builder logs')
+    # for log_group in imagebuilder_cw_log_list:
+    #     if delete_version_list_any_cloudwatch_log(delete_version_list, log_group):
+    #         cmd = ['logs', 'delete-log-group']
+    #         cmd += ['--log-group-name', log_group['logGroupName']]
+    #         aws_cli.run(cmd, ignore_error=True)
 
 
 ################################################################################
